refactor(structure): 拆分大型模块为子目录结构 + backon 迁移到调用方
- bootstrap: config.rs → config/, routes.rs → route_groups/, 新增 workers.rs
- config: 拆分 security/validation 为独立文件
- di: modules.rs → modules/ 子模块
- domain/services: llm_service.rs → llm/ 子模块, 新增 webhook_event_builder
- application: 拆分 crawl_use_case → state_machine + link_processor
- workers: 拆分 scrape_worker → executor + response_builder + link_extractor
- 所有 backoff 调用方迁移至 backon

+++ src/bootstrap/services.rs
@@ -231,7 +231,7 @@ pub async fn init_rate_limiting_service(
             strategy: ConcurrencyStrategy::DistributedSemaphore,
             max_concurrent_tasks: settings.concurrency.default_team_limit as u32,
             max_concurrent_per_team: settings.concurrency.default_team_limit as u32,
-            lock_timeout_seconds: settings.concurrency.task_lock_duration_seconds as u64,
+            lock_timeout_seconds: settings.concurrency.task_lock_duration_seconds,
             enabled: true,
         };
 
@@ -1012,7 +1012,7 @@ mod tests {
     // ========== init_team_semaphore tests ==========
 
     /// 构造测试用 Settings（ConcurrencySettings 自定义字段）
-    fn make_test_settings(default_team_limit: i64, adaptive_enabled: bool) -> Settings {
+    fn make_test_settings(default_team_limit: u64, adaptive_enabled: bool) -> Settings {
         let mut settings = Settings::default();
         settings.concurrency.default_team_limit = default_team_limit;
         settings.concurrency.adaptive_enabled = adaptive_enabled;
